Move popup to its own app
