+++ templates/core/badge/partial/badge_assignment.html
@@ -1,5 +1,5 @@
 {%include 'base/show_message.html'%}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {% comment %}
 	This template must be used with the popup system. To do so you need at least 2 things
