+++ templates/core/badge/partial/badge_create_form.html
@@ -1,4 +1,4 @@
-{% load static badge_tags custom_tags %}
+{% load static badge_tags custom_tags popup_tags %}
 {% load i18n %}
 
 {% block extra_css %}
@@ -75,8 +75,7 @@
                 <!-- ====================================================
                  Section 0 : Choisir si on importe une icon, ou on en génère une
 
-                 Section 0: Write title and subtitle.
-                 Each keystroke (300ms debounce) updates the preview.
+                 Section 0: Choose if we import an icon or if we generate one
                  ==================================================== -->
                 <h2 class="step-title mt-4">
                     <span class="step-number" aria-hidden="true">0</span>
