+++ templates/core/badge/partial/badge_edit_form.html
@@ -2,7 +2,7 @@
 {# Formulaire d'edition de badge pour la modale HTMX (popup). #}
 {# / Badge edit form for HTMX modal (popup). #}
 {% load i18n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 <h2>{% translate "Modifier le badge" %}</h2>
