+++ templates/core/badge/partial/badge_endorsement.html
@@ -1,5 +1,5 @@
 {%include 'base/show_message.html'%}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 <div id="badge-endorsment-form">
     <h2>Endosser un badge</h2>
