+++ templates/core/badge/partial/dream_badge_create.html
@@ -1,5 +1,5 @@
 {%include 'base/show_message.html'%}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 {% load i18n %}
 
 <div id="dream-badge-form">
