+++ templates/core/course/edit.html
@@ -5,7 +5,7 @@
 {% load i18n %}
 {% load static %}
 {% load l10n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {% block title %}
     {{ course.name }} — {% translate "Parcours" %} — openbadge.coop
