+++ templates/core/course/partial/add_badge_popup.html
@@ -1,5 +1,5 @@
 {%include 'base/show_message.html'%}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 
 <div>
