+++ templates/core/course/partial/badge_list_add.html
@@ -1,4 +1,4 @@
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 <style>
 /* Grille de sélection de badges - 4 par ligne
