+++ templates/core/course/partial/course_create.html
@@ -1,5 +1,5 @@
 {%include 'base/show_message.html'%}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 
 <div id="course-creation-form">
