+++ templates/core/home/partial/badge_focus.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {# Vue focus badge : le badge s'agrandit dans la colonne Badges, #}
 {# les deux autres colonnes filtrent sur les objets liés au badge. #}
