+++ templates/core/home/partial/multi_focus.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {# Vue multi-focus : 2 ou 3 items sélectionnés (cartes détail). #}
 {# Avec 2 items : la 3e colonne affiche l'intersection. #}
