+++ templates/core/home/partial/person_focus.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {# Vue focus personne : la personne s'agrandit dans la colonne Personnes, #}
 {# les deux autres colonnes filtrent sur les objets liés. #}
