+++ templates/core/home/partial/search_results.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {# Résultats de recherche en colonnes dynamiques : Badges / Structures / Personnes #}
 {# Search results in dynamic columns: Badges / Structures / People #}
