+++ templates/core/home/partial/structure_focus.html
@@ -1,5 +1,5 @@
 {% load i18n %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {# Vue focus structure : la structure s'agrandit dans la colonne Structures, #}
 {# les deux autres colonnes filtrent sur les objets liés. #}
