+++ templates/core/structure/partial/structure_edit_form.html
@@ -1,4 +1,4 @@
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 {% load l10n %}
 {% load static %}
