+++ templates/core/user/partial/user_profile_edit.html
@@ -1,4 +1,4 @@
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 <div id="user-info">
     <form hx-post="{% url 'core:edit-profile' user.pk %}"
