+++ templates/cytoscape/graph.html
@@ -1,4 +1,4 @@
-{% load static custom_tags %}
+{% load static custom_tags popup_tags %}
 
 <script src="{% static 'js/cytoscape.min.js' %}" defer></script> {# cytoscape js for the graph #}
 <script src="{% static 'js/dagre.min.js' %}" defer></script> {# Needed for cytoscape-dagre #}
