+++ templates/unused/core/assignments/detail.html
@@ -1,5 +1,5 @@
 {% extends 'base/base.html' %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {% block title %} Badge {{ assignment.badge.name }}{% endblock %}
