+++ templates/unused/core/assignments/list_user_assignment.html
@@ -1,5 +1,5 @@
 {% extends 'base/base.html' %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {% block title %}Badge {{ badge.name }} assigné à {{ user.username }}{% endblock %}
