+++ templates/unused/core/assignments/partials/assignment_list.html
@@ -1,4 +1,4 @@
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 
 <div class="badge-grid">
