+++ templates/unused/core/assignments/partials/assignment_list_by_badges.html
@@ -1,4 +1,4 @@
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 
 {% comment %}
