+++ templates/unused/core/users/cv_material.html
@@ -1,5 +1,5 @@
 {% extends 'base/base.html' %}
-{% load custom_tags %}
+{% load custom_tags popup_tags %}
 {% load humanize %}
 {% load custom_filters %}
